+++ pkg/evaluation/save.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/docker/docker-agent/pkg/chat"
 	"github.com/docker/docker-agent/pkg/session"
+	"github.com/docker/docker-agent/pkg/session/sqlitestore"
 	"github.com/docker/docker-agent/pkg/tools"
 )
 
@@ -27,7 +28,7 @@ func SaveRunSessions(ctx context.Context, run *EvalRun, outputDir string) (strin
 	}
 
 	// Create a new SQLite session store for this eval run
-	store, err := session.NewSQLiteSessionStore(ctx, dbPath)
+	store, err := sqlitestore.New(ctx, dbPath)
 	if err != nil {
 		return "", fmt.Errorf("creating session store: %w", err)
 	}
