+++ pkg/session/store_test.go
@@ -1,10 +1,8 @@
 package session
 
 import (
-	"context"
 	"database/sql"
 	"encoding/json"
-	"os"
 	"path/filepath"
 	"testing"
 	"time"
@@ -21,7 +19,7 @@ func TestStoreAgentName(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_store.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -75,7 +73,7 @@ func TestStoreMultipleAgents(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_store_multi.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -126,7 +124,7 @@ func TestGetSessions(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_get_sessions.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -176,7 +174,7 @@ func TestGetSessionSummaries(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_get_session_summaries.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -240,7 +238,7 @@ func TestGetSessionSummaries_CorruptCreatedAt(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_corrupt_created_at.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	sqlStore := store.(*SQLiteSessionStore)
 	defer sqlStore.Close()
@@ -292,7 +290,7 @@ func TestBranchSessionCopiesPrefix(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_branch_prefix.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -333,7 +331,7 @@ func TestBranchSessionClonesSubSession(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_branch_subsession.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -381,7 +379,7 @@ func TestStoreAgentNameJSON(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_store_json.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -418,32 +416,12 @@ func TestStoreAgentNameJSON(t *testing.T) {
 	assert.Equal(t, "another-agent", retrievedSession.Messages[2].Message.AgentName) // Second agent
 }
 
-func TestNewSQLiteSessionStore_DirectoryNotWritable(t *testing.T) {
-	t.Parallel()
-
-	readOnlyDir := filepath.Join(t.TempDir(), "readonly")
-	err := os.Mkdir(readOnlyDir, 0o555)
-	require.NoError(t, err)
-
-	_, err = NewSQLiteSessionStore(t.Context(), filepath.Join(readOnlyDir, "session.db"))
-	require.Error(t, err)
-
-	assert.Contains(t, err.Error(), "cannot create database")
-	assert.Contains(t, err.Error(), "permission denied or file cannot be created")
-
-	// We should surface the real "cannot create database" error directly instead of
-	// running the backup+retry recovery path (which cannot fix a filesystem-level
-	// problem and would only wrap the real error in a confusing "migration failed"
-	// message).
-	assert.NotContains(t, err.Error(), "migration failed")
-}
-
 func TestUpdateSession_LazyCreation(t *testing.T) {
 	t.Parallel()
 
 	tempDB := filepath.Join(t.TempDir(), "test_lazy.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -534,7 +512,7 @@ func TestStorePermissions(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_permissions.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -566,7 +544,7 @@ func TestStorePermissions_NilPermissions(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_nil_permissions.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -593,7 +571,7 @@ func TestUpdateSession_Permissions(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_update_permissions.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -629,7 +607,7 @@ func TestAgentModelOverrides_SQLite(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_model_overrides.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -664,7 +642,7 @@ func TestAgentModelOverrides_Update(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_model_overrides_update.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -700,7 +678,7 @@ func TestAgentModelOverrides_EmptyMap(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_model_overrides_empty.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -723,151 +701,6 @@ func TestAgentModelOverrides_EmptyMap(t *testing.T) {
 	assert.Empty(t, retrieved.AgentModelOverrides)
 }
 
-func TestNewSQLiteSessionStore_RejectsNewerDatabase(t *testing.T) {
-	t.Parallel()
-
-	tempDir := t.TempDir()
-	dbPath := filepath.Join(tempDir, "test_newer_db.db")
-
-	// Create a valid store first (applies all known migrations)
-	store, err := NewSQLiteSessionStore(t.Context(), dbPath)
-	require.NoError(t, err)
-	defer store.(*SQLiteSessionStore).Close()
-
-	// Inject a future migration into the database to simulate a newer version
-	db, err := sql.Open("sqlite", dbPath)
-	require.NoError(t, err)
-	_, err = db.ExecContext(t.Context(),
-		"INSERT INTO migrations (id, name, description, applied_at) VALUES (?, ?, ?, ?)",
-		9999, "9999_future_migration", "Added by a newer version", "2099-01-01T00:00:00Z")
-	require.NoError(t, err)
-	db.Close()
-
-	// Opening the store should fail with a clear error about version mismatch
-	_, err = NewSQLiteSessionStore(t.Context(), dbPath)
-	require.Error(t, err)
-	require.ErrorIs(t, err, ErrNewerDatabase)
-	assert.Contains(t, err.Error(), "9999")
-	assert.Contains(t, err.Error(), "upgrade docker-agent")
-}
-
-// TestNewSQLiteSessionStore_TransientErrorPreservesDB verifies that a
-// transient failure during open (here: a canceled context, e.g. Ctrl-C
-// during startup) does NOT trigger the backup-and-reset recovery path,
-// which would silently discard a healthy session history.
-func TestNewSQLiteSessionStore_TransientErrorPreservesDB(t *testing.T) {
-	t.Parallel()
-
-	dbPath := filepath.Join(t.TempDir(), "test_transient.db")
-
-	// Create a valid database with one session.
-	store, err := NewSQLiteSessionStore(t.Context(), dbPath)
-	require.NoError(t, err)
-	require.NoError(t, store.AddSession(t.Context(), &Session{ID: "keep-me", CreatedAt: time.Now()}))
-	require.NoError(t, store.Close())
-
-	// Opening with an already-canceled context must fail without recovery.
-	canceled, cancel := context.WithCancel(t.Context())
-	cancel()
-	_, err = NewSQLiteSessionStore(canceled, dbPath)
-	require.ErrorIs(t, err, context.Canceled)
-
-	// The database must be untouched: no .bak, data still there.
-	_, err = os.Stat(dbPath + ".bak")
-	assert.True(t, os.IsNotExist(err), "no backup should be created for transient errors")
-
-	store, err = NewSQLiteSessionStore(t.Context(), dbPath)
-	require.NoError(t, err)
-	defer store.Close()
-	retrieved, err := store.GetSession(t.Context(), "keep-me")
-	require.NoError(t, err)
-	assert.Equal(t, "keep-me", retrieved.ID)
-}
-
-func TestNewSQLiteSessionStore_MigrationFailureRecovery(t *testing.T) {
-	t.Parallel()
-
-	tempDir := t.TempDir()
-	dbPath := filepath.Join(tempDir, "test_migration_recovery.db")
-	backupPath := dbPath + ".bak"
-
-	// Create a corrupted database file that will fail migrations
-	err := os.WriteFile(dbPath, []byte("not a valid sqlite database"), 0o644)
-	require.NoError(t, err)
-
-	// Opening should trigger recovery: backup the corrupt file and create fresh db
-	store, err := NewSQLiteSessionStore(t.Context(), dbPath)
-	require.NoError(t, err)
-	defer store.(*SQLiteSessionStore).Close()
-
-	// Verify a backup was created
-	_, err = os.Stat(backupPath)
-	require.NoError(t, err, "backup file should exist")
-
-	// Verify the store works with the fresh database
-	session := &Session{
-		ID:        "test-session",
-		CreatedAt: time.Now(),
-	}
-	err = store.AddSession(t.Context(), session)
-	require.NoError(t, err)
-
-	retrieved, err := store.GetSession(t.Context(), "test-session")
-	require.NoError(t, err)
-	assert.Equal(t, "test-session", retrieved.ID)
-}
-
-func TestBackupDatabase(t *testing.T) {
-	t.Parallel()
-
-	t.Run("backs up existing database file", func(t *testing.T) {
-		tempDir := t.TempDir()
-		dbPath := filepath.Join(tempDir, "test.db")
-		backupPath := dbPath + ".bak"
-
-		// Create a file to backup
-		err := os.WriteFile(dbPath, []byte("test content"), 0o644)
-		require.NoError(t, err)
-
-		// Also create WAL and SHM files
-		err = os.WriteFile(dbPath+"-wal", []byte("wal content"), 0o644)
-		require.NoError(t, err)
-		err = os.WriteFile(dbPath+"-shm", []byte("shm content"), 0o644)
-		require.NoError(t, err)
-
-		// Backup the database
-		err = backupDatabase(dbPath)
-		require.NoError(t, err)
-
-		// Original should be gone
-		_, err = os.Stat(dbPath)
-		assert.True(t, os.IsNotExist(err), "original file should be moved")
-
-		// WAL and SHM should also be gone
-		_, err = os.Stat(dbPath + "-wal")
-		assert.True(t, os.IsNotExist(err), "WAL file should be moved")
-		_, err = os.Stat(dbPath + "-shm")
-		assert.True(t, os.IsNotExist(err), "SHM file should be moved")
-
-		// Check backup files exist
-		_, err = os.Stat(backupPath)
-		require.NoError(t, err, "main backup should exist")
-		_, err = os.Stat(backupPath + "-wal")
-		require.NoError(t, err, "WAL backup should exist")
-		_, err = os.Stat(backupPath + "-shm")
-		require.NoError(t, err, "SHM backup should exist")
-	})
-
-	t.Run("handles nonexistent file gracefully", func(t *testing.T) {
-		tempDir := t.TempDir()
-		dbPath := filepath.Join(tempDir, "nonexistent.db")
-
-		// Backup should succeed (nothing to backup)
-		err := backupDatabase(dbPath)
-		require.NoError(t, err)
-	})
-}
-
 // TestOrphanedSubsessionReference verifies that loading sessions gracefully
 // handles orphaned subsession references (where the subsession was deleted
 // but the reference in session_items remains).
@@ -876,7 +709,7 @@ func TestOrphanedSubsessionReference(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_orphaned.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -972,7 +805,7 @@ func TestMigration_ExistingMessagesToSessionItems(t *testing.T) {
 	db.Close()
 
 	// Now open with the store, which runs migrations
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -990,7 +823,7 @@ func TestAddError_SQLiteRoundTrip(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_add_error.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -1089,7 +922,7 @@ func TestSQLiteStore_CostsSurviveReload(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_cost_roundtrip.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 
 	root := New(WithID("cost-roundtrip-root"))
@@ -1110,7 +943,7 @@ func TestSQLiteStore_CostsSurviveReload(t *testing.T) {
 	require.InDelta(t, 0.18, want, 1e-9)
 	require.NoError(t, store.(*SQLiteSessionStore).Close())
 
-	reopened, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	reopened, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer reopened.(*SQLiteSessionStore).Close()
 
@@ -1175,7 +1008,7 @@ func TestResolveSessionID_SQLite(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_resolve.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
 
@@ -1234,7 +1067,7 @@ func TestResolveSessionID_ExcludesSubSessions(t *testing.T) {
 
 	tempDB := filepath.Join(t.TempDir(), "test_resolve_subsessions.db")
 
-	store, err := NewSQLiteSessionStore(t.Context(), tempDB)
+	store, err := newSQLiteStoreForTest(t, tempDB)
 	require.NoError(t, err)
 	defer store.(*SQLiteSessionStore).Close()
